Fixing deprecation and other warnings.

+++ bigtable-client-core-parent/bigtable-client-core/src/main/java/com/google/cloud/bigtable/config/BigtableOptions.java
@@ -65,7 +65,7 @@ public class BigtableOptions implements Serializable, Cloneable {
     // 20 Channels seemed to work well on a 4 CPU machine, and this ratio seems to scale well for
     // higher CPU machines. Use no more than 250 Channels by default.
     int availableProcessors = Runtime.getRuntime().availableProcessors();
-    return (int) Math.min(250, Math.max(1, availableProcessors * 4));
+    return Math.min(250, Math.max(1, availableProcessors * 4));
   }
 
   public static BigtableOptions getDefaultOptions() {
